K and S are not final, K equals has been changed
A way to declare that K is the same with T, K is the same with OR and T is the samew with OR. In java equality is expressed via the .hashcode() and equals(Obj) method. Therefore K and S are not final, so they can be extended (a decision deferred until now). Another solution would be to use another equality/equivalence model/method such as an additiona function. However the semantics of using a suitable equals are easier to understand. Furthermore S and K are designed as Singletons and are stateless.So there should exist only one instance of K and one instance of S. Even if this was not true, every S instance should be considered equal to every S instance. Similarly every K instance should be considered equal to every K instance.